refactor: restructure files and remove unused code
