--- src/command/mod.rs
@@ -1,5 +1,0 @@
-pub mod cmd;
-pub mod complete;
-pub mod init;
-pub mod prompt;
-pub mod query;
